% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exampleScores.R
\docType{data}
\name{exampleScores}
\alias{exampleScores}
\title{Example multivariate scores dataset}
\usage{
data(exampleScores)
}
\description{
Example arbitrary multivariate dataset consisting of three groups with arbitrary labels of 1, 2 and 3. The dataset was generated
for testing purposes. It contains two objects; a matrix \code{mat} containing the multivariate samples and a corresponding labeling
vector \code{lab}
}
\keyword{datasets}
