% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msiScores-package.r
\docType{package}
\name{msiScores}
\alias{msiScores}
\title{msiScores: Statistical & multivariate scores for systematic workflow standardization for Mass Spectrometry imaging (MSI) data.}
\description{
Statistical & multivariate scores implementation as described
in [1].  The scores can be used for
systematic evaluation of workflows and comparison of sample preperation
methods to check for reproducibility and reliability of the generated data from
different preparation protocols.
}
\references{
[1] Erich, Katrin, et al. "Scores for standardization of on-tissue digestion of formalin-fixed paraffin-embedded tissue in
MALDI-MS imaging." Biochimica et Biophysica Acta (BBA)-Proteins and Proteomics 1865.7 (2017): 907-915.
}
\author{
Denis A. Sammour, \email{d.abu-sammour@hs-mannheim.de}
}
