% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/natFC.R
\name{natFC}
\alias{natFC}
\title{Pixel-wise foldchange analysis}
\usage{
natFC(
  data.mat,
  sample.labels,
  n.samples = 3000,
  box.plot = TRUE,
  col.vec = NULL
)
}
\arguments{
\item{data.mat}{A matrix holding the intensity information of all spectra with rows having the samples (pixels or spectra) and columns
denoting variables (m/z features in case of MS).}

\item{sample.labels}{A character vector of length \code{nrow(data.matrix)} holding the labels of the samples (pixels or spectra).}

\item{n.samples}{The size of the sampling pool. Defaults to 3000.}

\item{box.plot}{A boolean used to plot the results in a boxplot. Defaults to TRUE.}

\item{col.vec}{A character vector specifying the colors of the groups with its length equal to number of unique groups in
\code{sample.labels}. If not provided, arbitrary colors will be assigned.}
}
\value{
returns a list of natFC values for each group indicated in the \code{sample.labels} object as well as overall natFC sampled from
all pixels of \code{data.mat}.
}
\description{
Performs foldchange analysis on randomly chosen, different combinations of observations of a given data matrix.
}
\details{
The function evaluates the so called natural foldchange on a paired randomly chosen combinations of observations sampled from
\code{data.mat}. For each pair of samples the function computes \code{natFD = max(|5th FC precentile|, |95th FC percentile|)}. The computed
values are then pooled into a list named based on the supplied groups as indicated in \code{sample.labels}.

For more details about the generated scores and for citation please refere to "Erich, Katrin, et al. Biochimica et Biophysica Acta (BBA)-
Proteins and Proteomics 1865.7 (2017): 907-915".
}
\author{
Denis Abu Sammour, \email{d.abu-sammour@hs-mannheim.de}
}
