% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcChart.R
\name{pcChart}
\alias{pcChart}
\title{Variances chart of PCA}
\usage{
pcChart(pca.object)
}
\arguments{
\item{pca.object}{The PCA list of class \code{prcomp}.}
}
\value{
returns a plot illustrating the variance explained by the principal components.
}
\description{
Computes the variances (cumulative and individual) per principal component and plots them.
}
\author{
Denis Abu Sammour, \email{d.abu-sammour@hs-mannheim.de}
}
