% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randR2.R
\name{randR2}
\alias{randR2}
\title{Pixel-wise correlation analysis}
\usage{
randR2(
  data.mat,
  sample.labels,
  n.samples = 3000,
  box.plot = TRUE,
  col.vec = NULL
)
}
\arguments{
\item{data.mat}{A matrix holding the intensity information of all spectra with rows having the samples (pixels or spectra) and columns
denoting variables (m/z features in case of MS).}

\item{sample.labels}{A character vector of length \code{nrow(data.matrix)} holding the labels of the samples (pixels or spectra).}

\item{n.samples}{The size of the sampling pool. Defaults to 3000.}

\item{box.plot}{A boolean used to plot the results in a boxplot. Defaults to TRUE.}

\item{col.vec}{A character vector specifying the colors of the groups with its length equal to number of unique groups in
\code{sample.labels}. If not provided, arbitrary colors will be assigned.}
}
\value{
returns a list of R2 values for each group indicated in the \code{sample.labels} object as well as overall R2 sampled from
all pixels of \code{data.mat}.
}
\description{
Computes the coefficient of determination based on randomized sub-sampling of the observations.
}
\details{
For more details about the generated scores and for citation please refere to "Erich, Katrin, et al. Biochimica et Biophysica Acta (BBA)-
Proteins and Proteomics 1865.7 (2017): 907-915".
}
\author{
Denis Abu Sammour, \email{d.abu-sammour@hs-mannheim.de}
}
